requirements.txt
# Core dependencies
openai>=1.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Optional dependencies  
redis>=5.0.4

# Visualization (existing)
matplotlib>=3.9.0
networkx>=3.2.0

# TensorBoard support
tensorboard>=2.15.0
torch>=2.0.0
torchvision>=0.15.0

# Additional useful libraries
typing-extensions>=4.5.0
email-validator>=2.0.0 